infrastructure/tests/requirements.txt
# Infrastructure Property-Based Testing Dependencies
pytest>=7.0.0
hypothesis>=6.0.0
python-hcl2>=4.0.0
pyyaml>=6.0.0
